Backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
idna==3.10
lxml==5.3.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
timedelta==2020.12.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
